+++ resources/views/markdown/changelog.md
@@ -7,6 +7,8 @@ Travis Michael Cross <a href="mailto:travis@not.tv">travis@not.tv</a>
 May 29, 2024
 
 * Fix how recurring scheduled dates are calculated, they need to be calculated in the timezone they were created in, then converted to UTC.
+* Fix the team members not showing up in the Team Manage page.
+* Scheduled a virus scan for 9am UTC.
 
 ## v0.9.3.3
 May 29, 2024